Extract NIP-07 bridge message handling into a hook.
Move permission prompting and bridge API calls out of BrowserScreen and
queue pending messages per origin in a Map.
